+++ src/typings/files.ts
@@ -1,1 +1,2 @@
-export type LinterAvailableFiles = 'ts' | 'tsx' | 'js' | 'jsx' | 'md' | 'json'
+export type LinterAvailableFiles = 'ts' | 'tsx' | 'js' | 'jsx'
+export type FormatAvailableFiles = 'md' | 'json' | LinterAvailableFiles